feat: audio. Now It's working. Two sounds, one for the launch and one for the explosion.
feat: the firework library is now implementd as a map, where the firework's ID is the key.

chore: Refactor the life of Firework objects. Now They live as *copies* inside event and shell objects. Living as copies inside events is necessary as we want multiple istances of library's fireworks. They are just templates, specifically available to be tweaked as the user prefers. I do not know if copies inside shell objects are really necessary.